+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-# 0.0.2
+# 0.0.3
+- token rename - 2 macros
+- smooth movement
+- rotate
+- end movement
+- new sound,animation